Modify email filter example to more generic one
This changes the email-filter example to a more generic one (example. com instead of a possibly existing domain name) that also enforces the domain part to be explicitly the one in the example and not merely containing the example somewhere. Before an email address of "mydomain. loc@example.com" would have been valid as well.